backend/requirements.txt
Django>=5.0,<6
psycopg[binary]
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
